Add jmp, jmpz, and stack offest ops
Some instructions to make loops possible. Had to rearrange the parser
a bit because of common prefixes.

Wound up bringing back the Assign instruction in a slightly different
way, and the LPush pseudo-instruction as a new instruction with
different behaviour.

Added just the bare minimum for the sum loop example- other jump
conditions to come.